+++ internal/handlers/handlers.go
@@ -6,9 +6,9 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/jtrpka0912/bookings.udemy.go/pkg/config"
-	"github.com/jtrpka0912/bookings.udemy.go/pkg/models"
-	"github.com/jtrpka0912/bookings.udemy.go/pkg/render"
+	"github.com/jtrpka0912/bookings.udemy.go/internal/config"
+	"github.com/jtrpka0912/bookings.udemy.go/internal/models"
+	"github.com/jtrpka0912/bookings.udemy.go/internal/render"
 )
 
 // Repo the repository used by the handlers
